refactor(webhook): nest addon validation package
Keep addon checks under Application ownership and stop tracking local superpowers specs.
